requirements.txt
adal==1.2.1
asn1crypto==0.24.0
azure-common==1.1.21
azure-mgmt-managementgroups==0.2.0
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
cryptography==2.7
idna==2.8
isodate==0.6.0
msrest==0.6.6
msrestazure==0.6.0
oauthlib==3.0.1
pycparser==2.19
PyJWT==1.7.1
python-dateutil==2.8.0
requests==2.22.0
requests-oauthlib==1.2.0
six==1.12.0
urllib3==1.25.3
